backend/requirements.txt
fastmcp==3.0.0b1
fastapi==0.128.0
uvicorn[standard]==0.40.0
openai==2.16.0
python-dotenv==1.2.1
httpx==0.28.1
slowapi==0.1.9
azure-cognitiveservices-speech==1.40.0
pytest==8.3.5
pytest-asyncio==0.25.2
pytest-mock==3.14.0
